refactor(jupiter): split the onboarding mega-hook into focused hooks
Decompose the 1,131-line useJupiterFlow hook into single-concern units,
no behavior change:
- useJupiterMessages: message stream, typing indicator, reveal helpers, transcript.
- useJupiterPublish: publish + bounded Google sync retry + sync-pending + nav.
- useGoogleCalendarOnboarding: connect redirect, calendar picker, schedule import.
- jupiterFlow.constants: storage keys, day/session maps, loadSaved + migration.
- useJupiterFlow now orchestrates step + answers state and the per-step handlers,
  composing the three sub-hooks. Storage keys re-exported for existing consumers.

Orchestrator 1131 to 542 lines. Lint + build clean, 30 tests pass.
